worker_base/worker/requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
psycopg2-binary
requests
pandas
urllib3